fixed format model form int to double

+++ app/src/main/java/com/imdglobal/psi/api/entities/LabelLocation.java
@@ -20,7 +20,7 @@ public double getLatitude() {
 return latitude;
 }
 
-public void setLatitude(int latitude) {
+public void setLatitude(double latitude) {
 this.latitude = latitude;
 }
 
@@ -28,7 +28,7 @@ public double getLongitude() {
 return longitude;
 }
 
-public void setLongitude(int longitude) {
+public void setLongitude(double longitude) {
 this.longitude = longitude;
 }
